requirements.txt
fastapi
pydantic
pydantic-settings
uvicorn
Pillow
python-magic
rasterio
python-multipart
pytest
httpx
